+++ c0.0.14a_08/src/phys/aabb.h
@@ -4,20 +4,20 @@
 #define AABB_H
 
 typedef struct {
-    double minX, minY, minZ;
-    double maxX, maxY, maxZ;
+    float minX, minY, minZ;
+    float maxX, maxY, maxZ;
 } AABB;
 
-AABB   AABB_create(double minX, double minY, double minZ, double maxX, double maxY, double maxZ);
-AABB   AABB_clone(const AABB* a);
-AABB   AABB_cloneMove(const AABB* a, double x, double y, double z);
-AABB   AABB_expand(const AABB* a, double x, double y, double z);
-AABB   AABB_grow(const AABB* a, double x, double y, double z);
-double AABB_clipXCollide(const AABB* a, const AABB* b, double x);
-double AABB_clipYCollide(const AABB* a, const AABB* b, double y);
-double AABB_clipZCollide(const AABB* a, const AABB* b, double z);
-int    AABB_intersects(const AABB* a, const AABB* b);
-void   AABB_move(AABB* a, double x, double y, double z);
-AABB   AABB_offset(const AABB* a, double x, double y, double z);
+AABB  AABB_create(float minX, float minY, float minZ, float maxX, float maxY, float maxZ);
+AABB  AABB_clone(const AABB* a);
+AABB  AABB_cloneMove(const AABB* a, float x, float y, float z);
+AABB  AABB_expand(const AABB* a, float x, float y, float z);
+AABB  AABB_grow(const AABB* a, float x, float y, float z);
+float AABB_clipXCollide(const AABB* a, const AABB* b, float x);
+float AABB_clipYCollide(const AABB* a, const AABB* b, float y);
+float AABB_clipZCollide(const AABB* a, const AABB* b, float z);
+int   AABB_intersects(const AABB* a, const AABB* b);
+void  AABB_move(AABB* a, float x, float y, float z);
+AABB  AABB_offset(const AABB* a, float x, float y, float z);
 
 #endif // AABB_H
